requirements.txt
asgiref==3.9.1
boto3==1.40.17
botocore==1.40.17
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.8.2
